+++ changelog.md
@@ -5,6 +5,7 @@
 - Prevent slug trait from overwriting the slug if provided on create
 - Validate slug when saving with dynamic slug trait
 - Set password reset email subject
+- Only display the params field on permissions that allow them on role form
 
 ## Version 2.3.3 (2017-02-13)
 - Add send password reset action to user listing